split mediawiki into pages & templates. Add content types
