+++ .changeset/little-comics-greet.md
@@ -1,0 +1,5 @@
+---
+"@osrs-wiki/cache-mediawiki": minor
+---
+
+Split mediawiki into pages and templates
