+++ .changeset/wise-swans-juggle.md
@@ -1,0 +1,5 @@
+---
+"@osrs-wiki/cache-mediawiki": minor
+---
+
+Create shared types for cache content
