+++ src/mediawiki/pages/combatAchievement/index.ts
@@ -1,2 +1,1 @@
-export * from "./combatAchievement.types";
 export { default as combatAchievementPageBuilder } from "./combatAchievement";
