+++ src/mediawiki/pages/item/item.test.ts
@@ -1,20 +1,20 @@
 import itemPageBuilder from "./item";
 import {
-  ATTACK_RANGE_PARAM,
-  ATTACK_SPEED_PARAM,
-  CRUSH_ATTACK_PARAM,
-  CRUSH_DEFENCE_PARAM,
-  MAGIC_ATTACK_PARAM,
-  MAGIC_DEFENCE_PARAM,
-  RANGED_ATTACK_PARAM,
-  RANGED_DEFENCE_PARAM,
-  SLASH_ATTACK_PARAM,
-  SLASH_DEFENCE_PARAM,
   STAB_ATTACK_PARAM,
+  SLASH_ATTACK_PARAM,
+  CRUSH_ATTACK_PARAM,
+  MAGIC_ATTACK_PARAM,
+  RANGED_ATTACK_PARAM,
+  ATTACK_SPEED_PARAM,
+  ATTACK_RANGE_PARAM,
+  SLASH_DEFENCE_PARAM,
+  CRUSH_DEFENCE_PARAM,
   STAB_DEFENCE_PARAM,
+  MAGIC_DEFENCE_PARAM,
+  RANGED_DEFENCE_PARAM,
   RANGED_AMMO_STRENGTH_PARAM,
   RANGED_EQUIPMENT_STRENGTH_PARAM,
-} from "./item.utils";
+} from "../../../types/item";
 import {
   CategoryID,
   Item,
