+++ src/mediawiki/pages/item/item.types.ts
@@ -1,49 +1,6 @@
 import { MediaWikiFile } from "@osrs-wiki/mediawiki-builder";
 
-export type WeaponSlot =
-  | "head"
-  | "weapon"
-  | "body"
-  | "legs"
-  | "shield"
-  | "cape"
-  | "2h"
-  | "hands"
-  | "feet"
-  | "neck"
-  | "ammot"
-  | "ring";
-
-export type WeaponType =
-  | "2h Sword"
-  | "Axe"
-  | "Banner"
-  | "Bladed Staff"
-  | "Blaster"
-  | "Bludgeon"
-  | "Blunt"
-  | "Bow"
-  | "Bulwark"
-  | "Chinchompas"
-  | "Claw"
-  | "Crossbow"
-  | "Gun"
-  | "Multi-Style"
-  | "Partisan"
-  | "Pickaxe"
-  | "Polearm"
-  | "Polestaff"
-  | "Powered Staff"
-  | "Salamander"
-  | "Scythe"
-  | "Slash Sword"
-  | "Spear"
-  | "Spiked"
-  | "Stab Sword"
-  | "Staff"
-  | "Thrown"
-  | "Whip"
-  | "Unarmed";
+import { WeaponSlot } from "../../../types/item/item";
 
 export type InfoboxBonuses = {
   astab: string;
