+++ src/mediawiki/templates/CombatStyles/index.ts
@@ -1,0 +1,1 @@
+export { default as CombatStyles } from "./CombatStyles";
