+++ src/mediawiki/templates/InfoboxBonuses/index.ts
@@ -1,0 +1,1 @@
+export { default as InfoboxBonuses } from "./InfoboxBonuses";
