+++ src/mediawiki/templates/InfoboxItem/InfoboxItem.test.ts
@@ -1,0 +1,66 @@
+import InfoboxItem from "./InfoboxItem";
+import { Item, ItemID, Params } from "../../../utils/cache2";
+
+const BASE_ITEM: Item = {
+  name: "Test Item",
+  isMembers: true,
+  isGrandExchangable: true,
+  id: 1234 as ItemID,
+  inventoryModel: undefined,
+  examine: "Examine text.",
+  zoom2d: 0,
+  xan2d: 0,
+  yan2d: 0,
+  offsetX2d: 0,
+  offsetY2d: 0,
+  isStackable: false,
+  price: 0,
+  wearpos1: undefined,
+  wearpos2: undefined,
+  wearpos3: undefined,
+  weight: 0,
+  maleModel: undefined,
+  maleOffset: 0,
+  maleModel1: undefined,
+  femaleModel: undefined,
+  femaleOffset: 0,
+  femaleModel1: undefined,
+  groundActions: [],
+  inventoryActions: ["Test Action"],
+  subops: [],
+  recolorFrom: [],
+  recolorTo: [],
+  retextureFrom: [],
+  retextureTo: [],
+  shiftClickIndex: 0,
+  maleModel2: undefined,
+  femaleModel2: undefined,
+  maleChatheadModel: undefined,
+  femaleChatheadModel: undefined,
+  maleChatheadModel2: undefined,
+  femaleChatheadModel2: undefined,
+  category: undefined,
+  zan2d: 0,
+  noteLinkedItem: undefined,
+  noteTemplate: undefined,
+  stackVariantItems: [],
+  stackVariantQuantities: [],
+  resizeX: 0,
+  resizeY: 0,
+  resizeZ: 0,
+  ambient: 0,
+  contrast: 0,
+  team: 0,
+  noted2: undefined,
+  noted3: undefined,
+  placeholderLinkedItem: undefined,
+  placeholderTemplate: undefined,
+  params: new Params(),
+};
+
+describe("IfnoboxItem", () => {
+  test("InfoboxItem - base", () => {
+    const item = InfoboxItem(BASE_ITEM);
+    expect(item).toMatchSnapshot();
+  });
+});
