+++ src/mediawiki/templates/InfoboxItem/index.ts
@@ -1,0 +1,1 @@
+export { default as InfoboxItem } from "./InfoboxItem";
