+++ src/mediawiki/templates/index.ts
@@ -1,0 +1,3 @@
+export * from "./CombatStyles";
+export * from "./InfoboxBonuses";
+export * from "./InfoboxItem";
