+++ src/types/item/item.ts
@@ -1,0 +1,46 @@
+import { Item, ParamID, WearPos } from "../../utils/cache2";
+
+export type WeaponSlot =
+  | "head"
+  | "weapon"
+  | "body"
+  | "legs"
+  | "shield"
+  | "cape"
+  | "2h"
+  | "hands"
+  | "feet"
+  | "neck"
+  | "ammot"
+  | "ring";
+
+export type WeaponType =
+  | "2h Sword"
+  | "Axe"
+  | "Banner"
+  | "Bladed Staff"
+  | "Blaster"
+  | "Bludgeon"
+  | "Blunt"
+  | "Bow"
+  | "Bulwark"
+  | "Chinchompas"
+  | "Claw"
+  | "Crossbow"
+  | "Gun"
+  | "Multi-Style"
+  | "Partisan"
+  | "Pickaxe"
+  | "Polearm"
+  | "Polestaff"
+  | "Powered Staff"
+  | "Salamander"
+  | "Scythe"
+  | "Slash Sword"
+  | "Spear"
+  | "Spiked"
+  | "Stab Sword"
+  | "Staff"
+  | "Thrown"
+  | "Whip"
+  | "Unarmed";
